+++ torch_geometric/datasets/lrgb.py
@@ -58,6 +58,8 @@ class LRGBDataset(InMemoryDataset):
             :obj:`torch_geometric.data.Data` object and returns a boolean
             value, indicating whether the data object should be included in the
             final dataset. (default: :obj:`None`)
+        force_reload (bool, optional): Whether to re-process the dataset.
+            (default: :obj:`False`)
 
     **STATS:**
 
@@ -131,14 +133,16 @@ class LRGBDataset(InMemoryDataset):
         transform: Optional[Callable] = None,
         pre_transform: Optional[Callable] = None,
         pre_filter: Optional[Callable] = None,
+        force_reload: bool = False,
     ):
         self.name = name.lower()
         assert self.name in self.names
         assert split in ['train', 'val', 'test']
 
-        super().__init__(root, transform, pre_transform, pre_filter)
+        super().__init__(root, transform, pre_transform, pre_filter,
+                         force_reload=force_reload)
         path = osp.join(self.processed_dir, f'{split}.pt')
-        self.data, self.slices = torch.load(path)
+        self.load(path)
 
     @property
     def raw_dir(self) -> str:
@@ -235,8 +239,8 @@ class LRGBDataset(InMemoryDataset):
 
                     data_list.append(data)
 
-                torch.save(self.collate(data_list),
-                           osp.join(self.processed_dir, f'{split}.pt'))
+                path = osp.join(self.processed_dir, f'{split}.pt')
+                self.save(data_list, path)
 
     def label_remap_coco(self):
         # Util function for name 'COCO-SP'
@@ -294,5 +298,4 @@ class LRGBDataset(InMemoryDataset):
 
                 data_list.append(data)
 
-            torch.save(self.collate(data_list),
-                       osp.join(self.processed_dir, f'{split}.pt'))
+            self.save(data_list, osp.join(self.processed_dir, f'{split}.pt'))