+++ torch_geometric/datasets/twitch.py
@@ -79,7 +79,7 @@ class Twitch(InMemoryDataset):
         self.name = name
         assert self.name in ['DE', 'EN', 'ES', 'FR', 'PT', 'RU']
         super().__init__(root, transform, pre_transform)
-        self.data, self.slices = torch.load(self.processed_paths[0])
+        self.load(self.processed_paths[0])
 
     @property
     def raw_dir(self) -> str:
@@ -113,4 +113,4 @@ class Twitch(InMemoryDataset):
         if self.pre_transform is not None:
             data = self.pre_transform(data)
 
-        torch.save(self.collate([data]), self.processed_paths[0])
+        self.save([data], self.processed_paths[0])