+++ torch_geometric/datasets/wikics.py
@@ -42,7 +42,7 @@ class WikiCS(InMemoryDataset):
             is_undirected = True
         self.is_undirected = is_undirected
         super().__init__(root, transform, pre_transform)
-        self.data, self.slices = torch.load(self.processed_paths[0])
+        self.load(self.processed_paths[0])
 
     @property
     def raw_file_names(self) -> List[str]:
@@ -87,4 +87,4 @@ class WikiCS(InMemoryDataset):
         if self.pre_transform is not None:
             data = self.pre_transform(data)
 
-        torch.save(self.collate([data]), self.processed_paths[0])
+        self.save([data], self.processed_paths[0])