+++ torch_geometric/transforms/generate_mesh_normals.py
@@ -9,7 +9,8 @@ from torch_geometric.utils import scatter
 @functional_transform('generate_mesh_normals')
 class GenerateMeshNormals(BaseTransform):
     r"""Generate normal vectors for each mesh node based on neighboring
-    faces (functional name: :obj:`generate_mesh_normals`)."""
+    faces (functional name: :obj:`generate_mesh_normals`).
+    """
     def forward(self, data: Data) -> Data:
         assert 'face' in data
         pos, face = data.pos, data.face